--- compiler/src/main/kotlin/plumb/internal/codegen/validator/Validator.kt
@@ -1,8 +1,0 @@
-package plumb.internal.codegen.validator
-
-import plumb.internal.codegen.Model
-import javax.lang.model.element.Element
-
-interface  Validator {
-    fun validate(element: Element, model: Model): Boolean
-}
