+++ compiler/src/main/kotlin/rxjoin/internal/codegen/step/ProcessStep.kt
@@ -1,0 +1,7 @@
+package rxjoin.internal.codegen.step
+
+import rxjoin.internal.codegen.Model
+
+interface ProcessStep {
+    fun process(model: Model)
+}
