+++ compiler/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,1 @@
-plumb.internal.codegen.PlumbProcessor
+rxjoin.internal.codegen.RxJoinProcessor
