--- core/src/main/java/plumb/Plumber.java
@@ -1,7 +1,0 @@
-package plumb;
-
-public interface Plumber<T1, T2> {
-    void plumb(T1 t1, T2 t2);
-
-    void demolish();
-}
