--- core/src/main/java/plumb/PlumberMap.java
@@ -1,7 +1,0 @@
-package plumb;
-
-public interface PlumberMap {
-    String IMPL_CLASS_FQCN = "plumb.PlumberMapImpl";
-
-    <T, R> Plumber<T, R> plumberFor(T t);
-}
