--- core/src/main/java/plumb/annotation/Out.java
@@ -1,12 +1,0 @@
-package plumb.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.SOURCE)
-@Target(ElementType.FIELD)
-public @interface Out {
-    String value();
-}
