--- core/src/main/java/plumb/annotation/Plumbed.java
@@ -1,12 +1,0 @@
-package plumb.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.SOURCE)
-@Target(ElementType.TYPE)
-public @interface Plumbed {
-    Class value();
-}
