+++ core/src/main/java/rxjoin/Joiner.java
@@ -1,0 +1,7 @@
+package rxjoin;
+
+public interface Joiner<T1, T2> {
+    void join(T1 t1, T2 t2);
+
+    void demolish();
+}
