+++ core/src/main/java/rxjoin/JoinerMap.java
@@ -1,0 +1,7 @@
+package rxjoin;
+
+public interface JoinerMap {
+    String IMPL_CLASS_FQCN = "rxjoin.JoinerMapImpl";
+
+    <T, R> Joiner<T, R> joinerFor(T t);
+}
