+++ core/src/main/java/rxjoin/annotation/In.java
@@ -1,0 +1,10 @@
+package rxjoin.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.SOURCE) @Target(ElementType.FIELD) public @interface In {
+    String value();
+}
