+++ rxjoin-example/src/main/java/rxjoin/example/activity/ColorPickerActivity.java
@@ -1,13 +1,12 @@
 package rxjoin.example.activity;
 
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
 import rxjoin.example.R;
 
 public class ColorPickerActivity extends AppCompatActivity {
 
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
+    @Override protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_color_picker);
     }
