+++ rxjoin-example/src/main/res/layout-land/activity_color_picker.xml
@@ -1,89 +1,101 @@
 <?xml version="1.0" encoding="utf-8"?>
 <rxjoin.example.view.ColorPickerView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:paddingBottom="@dimen/activity_vertical_margin"
-    android:paddingStart="@dimen/activity_horizontal_margin"
-    android:paddingEnd="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin">
+        xmlns:tools="http://schemas.android.com/tools"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:orientation="vertical"
+        android:paddingBottom="@dimen/activity_vertical_margin"
+        android:paddingEnd="@dimen/activity_horizontal_margin"
+        android:paddingStart="@dimen/activity_horizontal_margin"
+        android:paddingTop="@dimen/activity_vertical_margin"
+        >
 
     <ImageView
-        android:id="@+id/target"
-        android:layout_width="300dp"
-        android:layout_height="match_parent"
-        tools:src="#FFFF00" />
+            android:id="@+id/target"
+            android:layout_width="300dp"
+            android:layout_height="match_parent"
+            tools:src="#FFFF00"
+            />
 
     <RelativeLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_toEndOf="@id/target"
-        android:layout_centerVertical="true"
-        android:layout_marginStart="16dp">
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_centerVertical="true"
+            android:layout_marginStart="16dp"
+            android:layout_toEndOf="@id/target"
+            >
 
         <LinearLayout
-            android:id="@+id/red_container"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:gravity="center_vertical"
-            android:orientation="horizontal">
+                android:id="@+id/red_container"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:gravity="center_vertical"
+                android:orientation="horizontal"
+                >
 
             <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="R: " />
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:text="R: "
+                    />
 
             <SeekBar
-                android:id="@+id/red_seek_bar"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_weight="3"
-                android:max="255" />
+                    android:id="@+id/red_seek_bar"
+                    android:layout_width="0dp"
+                    android:layout_height="wrap_content"
+                    android:layout_weight="3"
+                    android:max="255"
+                    />
         </LinearLayout>
 
         <LinearLayout
-            android:id="@+id/green_container"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_below="@id/red_container"
-            android:layout_marginTop="32dp"
-            android:gravity="center_vertical"
-            android:orientation="horizontal">
+                android:id="@+id/green_container"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_below="@id/red_container"
+                android:layout_marginTop="32dp"
+                android:gravity="center_vertical"
+                android:orientation="horizontal"
+                >
 
             <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="G: " />
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:text="G: "
+                    />
 
             <SeekBar
-                android:id="@+id/green_seek_bar"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_weight="3"
-                android:max="255" />
+                    android:id="@+id/green_seek_bar"
+                    android:layout_width="0dp"
+                    android:layout_height="wrap_content"
+                    android:layout_weight="3"
+                    android:max="255"
+                    />
         </LinearLayout>
 
         <LinearLayout
-            android:id="@+id/blue_container"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_below="@id/green_container"
-            android:layout_marginTop="32dp"
-            android:gravity="center_vertical"
-            android:orientation="horizontal">
+                android:id="@+id/blue_container"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_below="@id/green_container"
+                android:layout_marginTop="32dp"
+                android:gravity="center_vertical"
+                android:orientation="horizontal"
+                >
 
             <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="B: " />
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:text="B: "
+                    />
 
             <SeekBar
-                android:id="@+id/blue_seek_bar"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_weight="3"
-                android:max="255" />
+                    android:id="@+id/blue_seek_bar"
+                    android:layout_width="0dp"
+                    android:layout_height="wrap_content"
+                    android:layout_weight="3"
+                    android:max="255"
+                    />
         </LinearLayout>
     </RelativeLayout>
